Factor `MiddlewareStream` out to new util.rs.
