+++ src/net/server/middleware/mod.rs
@@ -30,3 +30,4 @@ pub mod builder;
 pub mod chain;
 pub mod processor;
 pub mod processors;
+pub mod util;
